Реализация refresh-токена, коммит 1

+++ src/main/java/org/musicservice/demo/repository/user/VerificationTokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Integer> {
+public interface VerificationTokenRepository extends JpaRepository<VerificationToken, Long> {
 
     Optional<VerificationToken> findByToken(String token);
